Move to panther_default; set up default_test.py
